+++ app/src/main/java/com/dnd_9th_3_android/gooding/data/repository/GalleryRepository.kt
@@ -1,7 +1,7 @@
 package com.dnd_9th_3_android.gooding.data.repository
 
 import androidx.paging.PagingData
-import com.dnd_9th_3_android.gooding.data.model.GalleryImageData
+import com.dnd_9th_3_android.gooding.data.model.gallery.GalleryImageData
 import kotlinx.coroutines.flow.Flow
 
 interface GalleryRepository {
