+++ app/src/main/java/com/dnd_9th_3_android/gooding/data/repository/GalleryRepositoryImpl.kt
@@ -6,7 +6,7 @@ import androidx.paging.PagingData
 import androidx.paging.map
 import com.dnd_9th_3_android.gooding.data.GalleryLocalDataSource
 import com.dnd_9th_3_android.gooding.data.GalleryPagingSource
-import com.dnd_9th_3_android.gooding.data.model.GalleryImageData
+import com.dnd_9th_3_android.gooding.data.model.gallery.GalleryImageData
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.map
 import javax.inject.Inject
