+++ app/src/main/java/com/dnd_9th_3_android/gooding/data/repository/KakaoMapAddressRepository.kt
@@ -1,6 +1,6 @@
 package com.dnd_9th_3_android.gooding.data.repository
 
-import com.dnd_9th_3_android.gooding.data.model.KakaoMapData
+import com.dnd_9th_3_android.gooding.data.model.map.KakaoMapData
 
 interface KakaoMapAddressRepository {
     suspend fun getKakaoMapAddress(keyword: String): List<KakaoMapData>
