+++ app/src/main/java/com/dnd_9th_3_android/gooding/data/repository/KakaoMapAddressRepositoryImpl.kt
@@ -1,7 +1,7 @@
 package com.dnd_9th_3_android.gooding.data.repository
 
 import com.dnd_9th_3_android.gooding.data.KakaoMapAddressRemoteDataSource
-import com.dnd_9th_3_android.gooding.data.model.KakaoMapData
+import com.dnd_9th_3_android.gooding.data.model.map.KakaoMapData
 import javax.inject.Inject
 
 class KakaoMapAddressRepositoryImpl @Inject constructor(
