+++ app/src/main/java/com/dnd_9th_3_android/gooding/gallery/GalleryActivity.kt
@@ -1,39 +1,25 @@
-package com.dnd_9th_3_android.gooding
+package com.dnd_9th_3_android.gooding.gallery
 
 import android.content.Context
 import android.content.Intent
 import android.os.Bundle
 import android.util.Log
-import android.view.LayoutInflater
-import android.view.ViewGroup
 import android.widget.Toast
 import androidx.activity.viewModels
 import androidx.appcompat.app.AppCompatActivity
 import androidx.databinding.DataBindingUtil
 import androidx.fragment.app.Fragment
-import androidx.lifecycle.LiveData
-import androidx.lifecycle.MutableLiveData
-import androidx.lifecycle.ViewModel
 import androidx.lifecycle.lifecycleScope
-import androidx.lifecycle.viewModelScope
-import androidx.paging.PagingData
-import androidx.paging.PagingDataAdapter
-import androidx.paging.cachedIn
-import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.GridLayoutManager
 import androidx.recyclerview.widget.RecyclerView
-import com.dnd_9th_3_android.gooding.data.model.GalleryImageData
-import com.dnd_9th_3_android.gooding.data.repository.GalleryRepository
+import com.dnd_9th_3_android.gooding.FeedFragment
+import com.dnd_9th_3_android.gooding.MyGoodingFragment
+import com.dnd_9th_3_android.gooding.R
+import com.dnd_9th_3_android.gooding.record.Record01Activity
+import com.dnd_9th_3_android.gooding.data.model.gallery.GalleryImageData
 import com.dnd_9th_3_android.gooding.databinding.ActivityGalleryBinding
-import com.dnd_9th_3_android.gooding.databinding.ItemGalleryImageBinding
 import dagger.hilt.android.AndroidEntryPoint
-import dagger.hilt.android.lifecycle.HiltViewModel
-import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.StateFlow
-import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.launch
-import javax.inject.Inject
 
 @AndroidEntryPoint
 class GalleryActivity : AppCompatActivity() {
@@ -143,13 +129,15 @@ class GalleryActivity : AppCompatActivity() {
     }
 
     private fun initFragmentType(type: String): Fragment {
-        when (type) {
+        return when (type) {
             FEED_FRAGMENT -> {
-                return FeedFragment.newInstance()
+                FeedFragment.newInstance()
             }
+
             MY_GOODING_FRAGMENT -> {
-                return MyGoodingFragment.newInstance()
+                MyGoodingFragment.newInstance()
             }
+
             else -> {
                 throw IllegalArgumentException("There is no type: $type");
             }
