+++ app/src/main/java/com/dnd_9th_3_android/gooding/record/BottomSheetDatePicker.kt
@@ -1,17 +1,18 @@
-package com.dnd_9th_3_android.gooding
+package com.dnd_9th_3_android.gooding.record
 
 import android.graphics.drawable.GradientDrawable
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import com.dnd_9th_3_android.gooding.databinding.BottomSheetSelectCategoryBinding
+import com.dnd_9th_3_android.gooding.R
+import com.dnd_9th_3_android.gooding.databinding.BottomSheetDatePickerBinding
 import com.google.android.material.bottomsheet.BottomSheetDialogFragment
 import java.util.Calendar
 
-class BottomSheetSelectCategory: BottomSheetDialogFragment() {
-    private var _binding: BottomSheetSelectCategoryBinding? = null
-    private val binding: BottomSheetSelectCategoryBinding
+class BottomSheetDatePicker : BottomSheetDialogFragment() {
+    private var _binding: BottomSheetDatePickerBinding? = null
+    private val binding: BottomSheetDatePickerBinding
         get() = _binding!!
 
     override fun onCreateView(
@@ -19,7 +20,7 @@ class BottomSheetSelectCategory: BottomSheetDialogFragment() {
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
-        _binding = BottomSheetSelectCategoryBinding.inflate(inflater, container, false)
+        _binding = BottomSheetDatePickerBinding.inflate(inflater, container, false)
 
         setupView()
         setupListeners()
@@ -30,7 +31,7 @@ class BottomSheetSelectCategory: BottomSheetDialogFragment() {
     override fun getTheme(): Int = R.style.RounderBottomSheetDialog
 
     private fun setupView() {
-        binding.cvSelectCategory.background = GradientDrawable().apply {
+        binding.cvDatePicker.background = GradientDrawable().apply {
             val radius = resources.getDimension(R.dimen.bottom_sheet_radius)
             cornerRadii = floatArrayOf(radius, radius, radius, radius, 0f, 0f, 0f, 0f)
         }
@@ -42,14 +43,19 @@ class BottomSheetSelectCategory: BottomSheetDialogFragment() {
         }
 
         binding.btnComplete.setOnClickListener {
+            val currentTime = Calendar.getInstance()
+            val year = currentTime.get(Calendar.YEAR)
+            val month = currentTime.get(Calendar.MONTH)
+            val day = currentTime.get(Calendar.DAY_OF_MONTH)
+
 
             dismiss()
         }
     }
 
     companion object {
-        fun newInstance(): BottomSheetSelectCategory =
-            BottomSheetSelectCategory().apply {
+        fun newInstance(): BottomSheetDatePicker =
+            BottomSheetDatePicker().apply {
                 arguments = Bundle().apply {
 
                 }
